agent/requirements.txt
anthropic==0.40.0
python-dotenv==1.0.0
rich==13.7.0
flask==3.0.0
openai-whisper
